+++ datafusion/physical-expr/src/physical_expr.rs
@@ -19,7 +19,6 @@ use arrow::datatypes::{DataType, Schema};
 
 use arrow::record_batch::RecordBatch;
 
-use crate::PhysicalSortExpr;
 use datafusion_common::{ColumnStatistics, DataFusionError, Result, ScalarValue};
 use datafusion_expr::ColumnarValue;
 
@@ -137,147 +136,6 @@ impl PhysicalExprStats for BasicExpressionStats {
     }
 }
 
-/// a Trait for marking tree node types that are rewritable
-pub trait TreeNodeRewritable: Clone {
-    /// Transform the tree node using the given [TreeNodeRewriter]
-    /// It performs a depth first walk of an node and its children.
-    ///
-    /// For an node tree such as
-    /// ```text
-    /// ParentNode
-    ///    left: ChildNode1
-    ///    right: ChildNode2
-    /// ```
-    ///
-    /// The nodes are visited using the following order
-    /// ```text
-    /// pre_visit(ParentNode)
-    /// pre_visit(ChildNode1)
-    /// mutatate(ChildNode1)
-    /// pre_visit(ChildNode2)
-    /// mutate(ChildNode2)
-    /// mutate(ParentNode)
-    /// ```
-    ///
-    /// If an Err result is returned, recursion is stopped immediately
-    ///
-    /// If [`false`] is returned on a call to pre_visit, no
-    /// children of that node are visited, nor is mutate
-    /// called on that node
-    ///
-    fn transform_using<R: TreeNodeRewriter<Self>>(
-        self,
-        rewriter: &mut R,
-    ) -> Result<Self> {
-        let need_mutate = match rewriter.pre_visit(&self)? {
-            RewriteRecursion::Mutate => return rewriter.mutate(self),
-            RewriteRecursion::Stop => return Ok(self),
-            RewriteRecursion::Continue => true,
-            RewriteRecursion::Skip => false,
-        };
-
-        let after_op_children =
-            self.map_children(|node| node.transform_using(rewriter))?;
-
-        // now rewrite this node itself
-        if need_mutate {
-            rewriter.mutate(after_op_children)
-        } else {
-            Ok(after_op_children)
-        }
-    }
-
-    /// Convenience utils for writing optimizers rule: recursively apply the given `op` to the node tree.
-    /// When `op` does not apply to a given node, it is left uncshanged.
-    /// The default tree traversal direction is transform_up(Postorder Traversal).
-    fn transform<F>(self, op: &F) -> Result<Self>
-    where
-        F: Fn(Self) -> Option<Self>,
-    {
-        self.transform_up(op)
-    }
-
-    /// Convenience utils for writing optimizers rule: recursively apply the given 'op' to the node and all of its
-    /// children(Preorder Traversal).
-    /// When the `op` does not apply to a given node, it is left unchanged.
-    fn transform_down<F>(self, op: &F) -> Result<Self>
-    where
-        F: Fn(Self) -> Option<Self>,
-    {
-        let node_cloned = self.clone();
-        let after_op = match op(node_cloned) {
-            Some(value) => value,
-            None => self,
-        };
-        after_op.map_children(|node| node.transform_down(op))
-    }
-
-    /// Convenience utils for writing optimizers rule: recursively apply the given 'op' first to all of its
-    /// children and then itself(Postorder Traversal).
-    /// When the `op` does not apply to a given node, it is left unchanged.
-    fn transform_up<F>(self, op: &F) -> Result<Self>
-    where
-        F: Fn(Self) -> Option<Self>,
-    {
-        let after_op_children = self.map_children(|node| node.transform_up(op))?;
-
-        let after_op_children_clone = after_op_children.clone();
-        let new_node = match op(after_op_children) {
-            Some(value) => value,
-            None => after_op_children_clone,
-        };
-        Ok(new_node)
-    }
-
-    /// Apply transform `F` to the node's children, the transform `F` might have a direction(Preorder or Postorder)
-    fn map_children<F>(self, transform: F) -> Result<Self>
-    where
-        F: FnMut(Self) -> Result<Self>;
-}
-
-/// Trait for potentially recursively transform an [`TreeNodeRewritable`] node
-/// tree. When passed to `TreeNodeRewritable::transform_using`, `TreeNodeRewriter::mutate` is
-/// invoked recursively on all nodes of a tree.
-pub trait TreeNodeRewriter<N: TreeNodeRewritable>: Sized {
-    /// Invoked before (Preorder) any children of `node` are rewritten /
-    /// visited. Default implementation returns `Ok(RewriteRecursion::Continue)`
-    fn pre_visit(&mut self, _node: &N) -> Result<RewriteRecursion> {
-        Ok(RewriteRecursion::Continue)
-    }
-
-    /// Invoked after (Postorder) all children of `node` have been mutated and
-    /// returns a potentially modified node.
-    fn mutate(&mut self, node: N) -> Result<N>;
-}
-
-/// Controls how the [TreeNodeRewriter] recursion should proceed.
-#[allow(dead_code)]
-pub enum RewriteRecursion {
-    /// Continue rewrite / visit this node tree.
-    Continue,
-    /// Call 'op' immediately and return.
-    Mutate,
-    /// Do not rewrite / visit the children of this node.
-    Stop,
-    /// Keep recursive but skip apply op on this node
-    Skip,
-}
-
-impl TreeNodeRewritable for Arc<dyn PhysicalExpr> {
-    fn map_children<F>(self, transform: F) -> Result<Self>
-    where
-        F: FnMut(Self) -> Result<Self>,
-    {
-        if !self.children().is_empty() {
-            let new_children: Result<Vec<_>> =
-                self.children().into_iter().map(transform).collect();
-            with_new_children_if_necessary(self, new_children?)
-        } else {
-            Ok(self)
-        }
-    }
-}
-
 /// Returns a copy of this expr if we change any child according to the pointer comparison.
 /// The size of `children` must be equal to the size of `PhysicalExpr::children()`.
 /// Allow the vtable address comparisons for PhysicalExpr Trait Objects，it is harmless even
@@ -301,52 +159,6 @@ pub fn with_new_children_if_necessary(
     } else {
         Ok(expr)
     }
-}
-
-/// Compare the two expr lists are equal no matter the order.
-/// For example two InListExpr can be considered to be equals no matter the order:
-///
-/// In('a','b','c') == In('c','b','a')
-pub fn expr_list_eq_any_order(
-    list1: &[Arc<dyn PhysicalExpr>],
-    list2: &[Arc<dyn PhysicalExpr>],
-) -> bool {
-    if list1.len() == list2.len() {
-        let mut expr_vec1 = list1.to_vec();
-        let mut expr_vec2 = list2.to_vec();
-        while let Some(expr1) = expr_vec1.pop() {
-            if let Some(idx) = expr_vec2.iter().position(|expr2| expr1.eq(expr2)) {
-                expr_vec2.swap_remove(idx);
-            } else {
-                break;
-            }
-        }
-        expr_vec1.is_empty() && expr_vec2.is_empty()
-    } else {
-        false
-    }
-}
-
-/// Strictly compare the two expr lists are equal in the given order.
-#[allow(dead_code)]
-pub fn expr_list_eq_strict_order(
-    list1: &[Arc<dyn PhysicalExpr>],
-    list2: &[Arc<dyn PhysicalExpr>],
-) -> bool {
-    list1.len() == list2.len() && list1.iter().zip(list2.iter()).all(|(e1, e2)| e1.eq(e2))
-}
-
-/// Strictly compare the two sort expr lists in the given order.
-///
-/// For Physical Sort Exprs, the order matters:
-///
-/// SortExpr('a','b','c') != SortExpr('c','b','a')
-#[allow(dead_code)]
-pub fn sort_expr_list_eq_strict_order(
-    list1: &[PhysicalSortExpr],
-    list2: &[PhysicalSortExpr],
-) -> bool {
-    list1.len() == list2.len() && list1.iter().zip(list2.iter()).all(|(e1, e2)| e1.eq(e2))
 }
 
 pub fn down_cast_any_ref(any: &dyn Any) -> &dyn Any {
@@ -411,10 +223,7 @@ mod tests {
     use std::sync::Arc;
 
     use super::*;
-    use crate::expressions::Column;
-    use crate::PhysicalSortExpr;
     use arrow::array::Int32Array;
-    use arrow::compute::SortOptions;
     use datafusion_common::Result;
 
     #[test]
@@ -506,133 +315,6 @@ mod tests {
         let no_boundaries =
             ExprBoundaries::new(ScalarValue::Int32(None), ScalarValue::Int32(None), None);
         assert_eq!(no_boundaries.reduce(), Some(ScalarValue::Int32(None)));
-
-        Ok(())
-    }
-
-    #[test]
-    fn expr_list_eq_any_order_test() -> Result<()> {
-        let list1: Vec<Arc<dyn PhysicalExpr>> = vec![
-            Arc::new(Column::new("a", 0)),
-            Arc::new(Column::new("a", 0)),
-            Arc::new(Column::new("b", 1)),
-        ];
-        let list2: Vec<Arc<dyn PhysicalExpr>> = vec![
-            Arc::new(Column::new("b", 1)),
-            Arc::new(Column::new("b", 1)),
-            Arc::new(Column::new("a", 0)),
-        ];
-        assert!(!expr_list_eq_any_order(list1.as_slice(), list2.as_slice()));
-        assert!(!expr_list_eq_any_order(list2.as_slice(), list1.as_slice()));
-
-        let list3: Vec<Arc<dyn PhysicalExpr>> = vec![
-            Arc::new(Column::new("a", 0)),
-            Arc::new(Column::new("b", 1)),
-            Arc::new(Column::new("c", 2)),
-            Arc::new(Column::new("a", 0)),
-            Arc::new(Column::new("b", 1)),
-        ];
-        let list4: Vec<Arc<dyn PhysicalExpr>> = vec![
-            Arc::new(Column::new("b", 1)),
-            Arc::new(Column::new("b", 1)),
-            Arc::new(Column::new("a", 0)),
-            Arc::new(Column::new("c", 2)),
-            Arc::new(Column::new("a", 0)),
-        ];
-        assert!(expr_list_eq_any_order(list3.as_slice(), list4.as_slice()));
-        assert!(expr_list_eq_any_order(list4.as_slice(), list3.as_slice()));
-        assert!(expr_list_eq_any_order(list3.as_slice(), list3.as_slice()));
-        assert!(expr_list_eq_any_order(list4.as_slice(), list4.as_slice()));
-
-        Ok(())
-    }
-
-    #[test]
-    fn sort_expr_list_eq_strict_order_test() -> Result<()> {
-        let list1: Vec<PhysicalSortExpr> = vec![
-            PhysicalSortExpr {
-                expr: Arc::new(Column::new("a", 0)),
-                options: SortOptions::default(),
-            },
-            PhysicalSortExpr {
-                expr: Arc::new(Column::new("a", 0)),
-                options: SortOptions::default(),
-            },
-            PhysicalSortExpr {
-                expr: Arc::new(Column::new("b", 1)),
-                options: SortOptions::default(),
-            },
-        ];
-
-        let list2: Vec<PhysicalSortExpr> = vec![
-            PhysicalSortExpr {
-                expr: Arc::new(Column::new("b", 1)),
-                options: SortOptions::default(),
-            },
-            PhysicalSortExpr {
-                expr: Arc::new(Column::new("a", 0)),
-                options: SortOptions::default(),
-            },
-            PhysicalSortExpr {
-                expr: Arc::new(Column::new("a", 0)),
-                options: SortOptions::default(),
-            },
-        ];
-
-        assert!(!sort_expr_list_eq_strict_order(
-            list1.as_slice(),
-            list2.as_slice()
-        ));
-        assert!(!sort_expr_list_eq_strict_order(
-            list2.as_slice(),
-            list1.as_slice()
-        ));
-
-        let list3: Vec<PhysicalSortExpr> = vec![
-            PhysicalSortExpr {
-                expr: Arc::new(Column::new("a", 0)),
-                options: SortOptions::default(),
-            },
-            PhysicalSortExpr {
-                expr: Arc::new(Column::new("b", 1)),
-                options: SortOptions::default(),
-            },
-            PhysicalSortExpr {
-                expr: Arc::new(Column::new("c", 2)),
-                options: SortOptions::default(),
-            },
-        ];
-        let list4: Vec<PhysicalSortExpr> = vec![
-            PhysicalSortExpr {
-                expr: Arc::new(Column::new("a", 0)),
-                options: SortOptions::default(),
-            },
-            PhysicalSortExpr {
-                expr: Arc::new(Column::new("b", 1)),
-                options: SortOptions::default(),
-            },
-            PhysicalSortExpr {
-                expr: Arc::new(Column::new("c", 2)),
-                options: SortOptions::default(),
-            },
-        ];
-
-        assert!(sort_expr_list_eq_strict_order(
-            list3.as_slice(),
-            list4.as_slice()
-        ));
-        assert!(sort_expr_list_eq_strict_order(
-            list4.as_slice(),
-            list3.as_slice()
-        ));
-        assert!(sort_expr_list_eq_strict_order(
-            list3.as_slice(),
-            list3.as_slice()
-        ));
-        assert!(sort_expr_list_eq_strict_order(
-            list4.as_slice(),
-            list4.as_slice()
-        ));
 
         Ok(())
     }
